Minor refactorings.
